+++ BlindBox Backend/src/main/java/com/blindbox/repository/BlindboxImageRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BlindboxImageRepository extends JpaRepository<BlindboxImage, Long> {
-    List<BlindboxImage> findByBlindbox_BlindboxID(Long blindboxId);
+public interface BlindboxImageRepository extends JpaRepository<BlindboxImage, Integer> {
+    List<BlindboxImage> findByBlindbox_BlindboxID(Integer blindboxId);
 }
