+++ BlindBox Backend/src/main/java/com/blindbox/repository/OrderDetailRepository.java
@@ -3,6 +3,6 @@ package com.blindbox.repository;
 import com.blindbox.model.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
     // Bạn có thể thêm các truy vấn tùy chỉnh nếu cần
 }
