+++ BlindBox Backend/src/main/java/com/blindbox/repository/ProductImageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
-    List<ProductImage> findByProductId(Long productId);
+public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+    List<ProductImage> findByProductId(Integer productId);
 }
