+++ BlindBox Backend/src/main/java/com/blindbox/repository/WishlistRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface WishlistRepository extends JpaRepository<Wishlist, Long> {
-    List<Wishlist> findByUserId(Long userId);
+public interface WishlistRepository extends JpaRepository<Wishlist, Integer> {
+    List<Wishlist> findByUserId(Integer userId);
 }
