+++ BlindBox Backend/src/main/java/com/blindbox/service/DiscountService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface DiscountService {
     List<Discount> getAllDiscounts();
-    Discount getDiscountById(Long id);
+    Discount getDiscountById(Integer id);
     Discount createDiscount(Discount discount);
-    Discount updateDiscount(Long id, Discount discount);
-    void deleteDiscount(Long id);
+    Discount updateDiscount(Integer id, Discount discount);
+    void deleteDiscount(Integer id);
 }
