+++ BlindBox Backend/src/main/java/com/blindbox/service/ProductService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getAllProducts();
-    Product getProductById(Long id);
+    Product getProductById(Integer id);
     Product createProduct(Product product);
-    Product updateProduct(Long id, Product product);
-    void deleteProduct(Long id);
+    Product updateProduct(Integer id, Product product);
+    void deleteProduct(Integer id);
 }
