+++ BlindBox Backend/src/main/java/com/blindbox/service/ReviewService.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 public interface ReviewService {
     List<Review> getAllReviews();
-    List<Review> getReviewsByUserId(Long userId);
-    List<Review> getReviewsByBlinkboxId(Long blinkboxId);
-    List<Review> getReviewsByProductId(Long productId);
-    Review getReviewById(Long id);
+    List<Review> getReviewsByUserId(Integer userId);
+    List<Review> getReviewsByBlinkboxId(Integer blinkboxId);
+    List<Review> getReviewsByProductId(Integer productId);
+    Review getReviewById(Integer id);
     Review createReview(Review review);
-    void deleteReview(Long id);
+    void deleteReview(Integer id);
 }
